+++ R/utils.R
@@ -66,7 +66,6 @@ equal_with_na <- function(a,b){
 #'
 #' @return a logical data.frame
 #'
-#' @examples
 df_compare <- function(x, y, verbose = TRUE) {
   compare_lgl <- purrr::map2_df(x, y, equal_with_na)
 
@@ -111,4 +110,43 @@ compare_init_post <- function(init, post_trait){
              select(post_trait, -dplyr::all_of(mismatch_col)))
 
 
+}
+
+is.evadata <- function(x) inherits(x, "evadata")
+
+#' An improved version of base R which.min that returns NA is all are Nas (instead of empty integer)
+#'
+#' @param x numeric vector whose min is searched for
+#'
+#' @return an integer vector with position of minimum value
+which_min <- function(x){
+  if(all(is.na(x)))
+    NA_integer_
+  else
+    which.min(x)
+}
+
+
+#' Column binds list of data.frames element by element
+#'
+#' This function will look at each element of the list y name (not by position) and will cbind the one with the same names
+#' It returns one list with names of all the input lists
+#'
+#' @param ... several named lists of data.frames
+#'
+#' @return a list of data.frames
+#'
+bind_list_df <- function(...){
+  lst_df <- list(...)
+
+  df_names <- purrr::map(lst_df, names) %>%
+    unlist() %>%
+    unique()
+
+  purrr::map(df_names,
+             .f = ~ dplyr::bind_cols(
+               purrr::map(lst_df,
+                          function(x) x[[.x]]))
+  ) %>%
+    stats::setNames(df_names)
 }