+++ tests/testthat/test-calc_distance.R
@@ -1,0 +1,4 @@
+test_that("calc_distance works", {
+  expect_type(calc_distance(evavelo_example),
+                  "list")
+})